Handle SKIPPED state better
The SKIPPED results have a fake URL and no run-time.

Make the run-time part of the match optional so we start matching
these lines.

Because the URL is fake, null the result for SKIPPED matches so the
job name won't be a link to nowhere.

Put a secondary-color span around the job name in the null case so
it's a bit more obvious it was skipped.

A weird thing with the CSS matchers is that "first-child" doesn't work
with the templating "dom-if" we use.  "first-child" seems to refer to
the "if={{job.link}}" template tag -- so it works when job.link is
defined, but when it is not, the "if={{!job.link}}" *isn't* the first
child, despite being the only thing displayed.  Anyway,
"first-of-type" seems more like what we want -- the first <td> element
no matter what ... that should always be the first column.  Update
that match.

Change-Id: I5c3bed85d7f64cd0890757237f93d13406c33788