Enable the plugin by default
New users who install the plugin are most likely going to expect
it to be enabled and work out-of-the-box.  Rather than ask users
to then also enable it by configuration, restore the default to
the previous behavior of automatically enabling it.

Projects can still disable this by default (or, presumably an admin
could set the default in All-Projects to false and then projects
can individually enable it).

Also, reword some documentation regarding Zuul's name.

Change-Id: I03b58bb6fa8b6db62793c05a0ec17309165c16f1