Switch to gr-icon from iron-icon
Gerrit 3.8 has largely deprecated iron-icons except for a small number
of cases which don't include the icons we are using in this plugin.
Switch to the material symbols font family provied through gr-icon which
appears to be the proper way to use check and close icons in Gerrit now.

This fixes missing icons under Gerrit 3.8. Gerrit 3.7 still renders them
using the old iron-icons but is expected to be compatibile with gr-icon
as well.

Change-Id: I96ffec351ac45293717b54fef73ef97bff80659c